Correct includes value in library.properties
The includes field in library.properties is used to customize which #include directives will be added to the sketch by the Arduino IDE via Sketch > Include Library > esp8266-wifi-picker. For this reason, the previous includes value was incorrect.

The includes field is only supported by Arduino IDE 1.6.10 and newer and those IDE versions do not require you to add #include directives for library dependencies so there is no need to have FS.h,ESP8266WiFi.h,WiFiClient.h,DNSServer.h,map in the includes field.

What is needed in the includes field is WiFiPicker.h, which I have added in lieu of the unnecessary file names.